Move ShareQrCodeModule to appropriate package
