+++ app/src/main/java/com/tari/android/wallet/ui/dialog/modular/modules/shareQr/ShareQRCodeModuleView.kt
@@ -1,4 +1,4 @@
-package com.tari.android.wallet.ui.screen.send.shareQr
+package com.tari.android.wallet.ui.dialog.modular.modules.shareQr
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -8,8 +8,8 @@ import com.tari.android.wallet.R
 import com.tari.android.wallet.databinding.DialogModuleShareQrCodeBinding
 import com.tari.android.wallet.ui.common.CommonViewModel
 import com.tari.android.wallet.ui.component.common.CommonView
-import com.tari.android.wallet.util.extension.dimenPx
 import com.tari.android.wallet.util.QrUtil
+import com.tari.android.wallet.util.extension.dimenPx
 
 @SuppressLint("ViewConstructor")
 class ShareQRCodeModuleView(context: Context, buttonModule: ShareQrCodeModule) :
